Rename a few source files
